Further refactor of ui componenet to element
