+++ src/main/java/ui/element/Button.java
@@ -1,4 +1,4 @@
-package ui.component;
+package ui.element;
 
 import ui.Element;
