+++ src/main/java/ui/element/ProgressBar.java
@@ -1,4 +1,4 @@
-package ui.component;
+package ui.element;
 
 import ui.Alignment;
 import ui.Element;
