+++ src/main/java/ui/element/TextHolder.java
@@ -1,4 +1,4 @@
-package ui.component;
+package ui.element;
 
 /**
  * @author Juyas
